Authorize Lesson 35 finished successfuly

+++ app/Http/Controllers/PostController.php
@@ -6,9 +6,9 @@ use App\Http\Requests\StorePostRequest;
 use App\Models\Category;
 use App\Models\Post;
 use App\Models\Tag;
-use Illuminate\Routing\Controllers\HasMiddleware;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
@@ -72,11 +72,14 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        Gate::authorize('update', $post);
+
         return view('posts.edit')->with(['post' => $post]);
     }
 
     public function update(StorePostRequest $request, Post $post)
     {
+
         if ($request->hasFile('photo')) {
 
             if (isset($post->photo)) {
@@ -100,6 +103,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        Gate::authorize('update', $post);
+
         if (isset($post->photo)) {
             Storage::delete($post->photo);
         }
